displays date next to title #106
