+++ client/styles/components/_record.scss
@@ -23,6 +23,11 @@
     }
   }
 
+  &__date {
+    font-size: 2rem;
+    font-weight: 500;
+  }
+
   &__dl {
     font-size: 1.25rem;
