+++ lib/transforms/get-values.js
@@ -32,5 +32,22 @@ module.exports = {
 
   getTitle (item) {
     return getPrimaryValue(item.attributes.title) || item.attributes.summary_title;
+  },
+
+  getDate (item) {
+    if (item.type === 'objects' || item.type === 'documents') {
+      return this.getMadeDate(item);
+    }
+
+    const birth = getNestedProperty(item, 'attributes.lifecycle.birth.0.date.0.latest') || null;
+    const death = getNestedProperty(item, 'attributes.lifecycle.death.0.date.0.latest') || null;
+
+    if (birth && death) {
+      return birth + ' - ' + death;
+    } else if (!death) {
+      return birth;
+    } else {
+      return 'Unknown - ' + death;
+    }
   }
 };
