+++ lib/transforms/json-to-html-data.js
@@ -11,8 +11,9 @@ module.exports = (resource) => {
   const details = getDetails(resource);
   const displayLocation = getValues.getDisplayLocation(resource.data);
   const level = getValues.getDocumentLevel(resource.data);
+  const date = getValues.getDate(resource.data);
 
-  return {title, type, fact, related, description, details, displayLocation, system, level};
+  return {title, type, fact, related, description, details, displayLocation, system, level, date};
 };
 
 function getType (data) {
@@ -25,12 +26,10 @@ function getFacts (resource) {
   const occupation = getOccupation(resource.data);
   const nationality = getNationality(resource.data);
   const born = getBorn(resource.data);
-  const birth = getBirth(resource.data);
-  const death = getDeath(resource.data);
   const makers = getMakers(resource);
   const made = getMade(resource);
 
-  return [occupation, nationality, born, birth, death, made].concat(makers).filter(el => el);
+  return [occupation, nationality, born, made].concat(makers).filter(el => el);
 }
 
 function getOccupation (data) {
@@ -58,16 +57,6 @@ function getBorn (data) {
   }
   const value = getNestedProperty(data, 'attributes.lifecycle.birth.0.location.name.0.value');
   return value ? {key: key, value: value} : null;
-}
-
-function getBirth (data) {
-  const value = getNestedProperty(data, 'attributes.lifecycle.birth.0.date.0.value');
-  return value ? {key: 'Birth', value: value} : null;
-}
-
-function getDeath (data) {
-  const value = getNestedProperty(data, 'attributes.lifecycle.death.0.date.0.value');
-  return value ? {key: 'Death', value: value} : null;
 }
 
 function getRelated (resource) {
