+++ lib/transforms/search-results-to-template-data.js
@@ -2,7 +2,6 @@ const QueryString = require('querystring');
 const formatNumber = require('format-number')();
 const Pagination = require('../pagination');
 const getValues = require('./get-values');
-const getNestedProperty = require('../nested-property');
 const getPrimaryValue = require('../get-primary-value');
 
 module.exports = (params, results, config) => {
@@ -48,7 +47,7 @@ function createResult (item) {
     type: item.type,
     title: getValues.getTitle(item),
     category: getCategory(item),
-    date: getDate(item),
+    date: getValues.getDate(item),
     figure: getFigure(item),
     figcaption: getFigCaption(item),
     link: item.links && item.links.self,
@@ -64,23 +63,6 @@ function getCategory (item) {
 
 function getOccupation (item) {
   return getPrimaryValue(item.attributes.occupation);
-}
-
-function getDate (item) {
-  if (item.type === 'objects' || item.type === 'documents') {
-    return getValues.getMadeDate(item);
-  }
-
-  const birth = getNestedProperty(item, 'attributes.lifecycle.birth.0.date.0.latest') || null;
-  const death = getNestedProperty(item, 'attributes.lifecycle.death.0.date.0.latest') || null;
-
-  if (birth && death) {
-    return birth + ' - ' + death;
-  } else if (!death) {
-    return birth;
-  } else {
-    return 'Unknown - ' + death;
-  }
 }
 
 function getFigure (item) {
