+++ templates/partials/records/record-top.html
@@ -1,7 +1,7 @@
   <section class="record-top {{class}}">
     <div class="row">
       <div class="large-12 columns">
-      <h1 class="record-top__title">{{#if level}}{{> global/icon i=level}}{{else}}{{> global/icon i=type}}{{/if}} {{title}}</h1>
+      <h1 class="record-top__title">{{#if level}}{{> global/icon i=level}}{{else}}{{> global/icon i=type}}{{/if}} {{title}} <span class="record-top__date">{{date}}</span></h1>
       {{#each fact}}
         <dl class="record-top__dl fact-{{classname this.key}}">
           <dt>{{ this.key }}:</dt>
